Split help function test
