+++ help_function/help_function.xml
@@ -193,6 +193,10 @@
     <test testName="sparse_utils_1" configFile="config/TEMPLATE_help_function_skip_cuda_backend.xml" />
     <test testName="sparse_utils_2_buffer" configFile="config/TEMPLATE_help_function_sparse_buffer.xml" splitGroup="double" />
     <test testName="sparse_utils_2_usm" configFile="config/TEMPLATE_help_function_sparse_usm.xml" splitGroup="double" />
+    <test testName="sparse_utils_3_buffer" configFile="config/TEMPLATE_help_function_sparse_buffer.xml" splitGroup="double" />
+    <test testName="sparse_utils_3_usm" configFile="config/TEMPLATE_help_function_sparse_usm.xml" splitGroup="double" />
+    <test testName="sparse_utils_4_buffer" configFile="config/TEMPLATE_help_function_sparse_buffer.xml" splitGroup="double" />
+    <test testName="sparse_utils_4_usm" configFile="config/TEMPLATE_help_function_sparse_usm.xml" splitGroup="double" />
     <test testName="atomic_compare_exchange_strong" configFile="config/TEMPLATE_help_function.xml" />
   </tests>
 </suite>
